Adjust up key handling.
If the up key is pressed on the input field the selectedIndex becomes
increasingly negative, hence the down key must then be pressed an equal
number of times (to "cancel" the up keypress).

This could be fixed by tested `selectedIndex < 0`, as well as the
existing test for `null`.  However, instead, change up to wrap to list
end, consistent with down wrapping to the top and helpful for long
lists.