Rename variables; use the selected element directly and remove offset.
Change several variable names to make clearer the values which are
JavaScript data and those which are DOM elements.

Use the `n-child()` selector to find the selected element directly,
after the `selectedIndex` has been changed, instead of using the first
result of find-by-class.

`isVisibleInside()` used `offset` to "look ahead or behind", from the
unchanged selected element, to find the "will be selected after
movement" element and then determine if it will be contained in the
parent.  However, as a result of working with the selected element
directly, `offset` can be removed: `isVisibleInside()` has the element
required for consideration.